UHF-7334: Support for OPENING_HOUR_OBJECT connections
