+++ src/Field/Connection/Highlight.php
@@ -5,6 +5,7 @@ declare(strict_types = 1);
 namespace Drupal\helfi_tpr\Field\Connection;
 
 use Drupal\Component\Utility\Html;
+use Drupal\Core\Field\FieldItemListInterface;
 
 /**
  * Provides a domain object for TPR connection type of HIGHLIGHT.
