+++ src/Field/Connection/OpeningHour.php
@@ -4,13 +4,10 @@ declare(strict_types = 1);
 
 namespace Drupal\helfi_tpr\Field\Connection;
 
-use Drupal\Component\Utility\Html;
-use Drupal\Core\Url;
-
 /**
- * Provides a domain object for TPR connection type of OPENING_HOURS.
+ * Provides a DTO for TPR connection type of OPENING_HOURS.
  */
-final class OpeningHour extends Connection {
+final class OpeningHour extends OpeningHourBase {
 
   /**
    * The type name.
@@ -18,44 +15,5 @@ final class OpeningHour extends Connection {
    * @var string
    */
   public const TYPE_NAME = 'OPENING_HOURS';
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getFields(): array {
-    return [
-      'name',
-      'www',
-    ];
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function build(): array {
-    $markup = Html::escape($this->get('name'));
-
-    if (function_exists('_filter_autop')) {
-      $markup = _filter_autop($markup);
-    }
-
-    $build = [
-      'name' => [
-        '#markup' => $markup,
-      ],
-    ];
-
-    if ($link = $this->get('www')) {
-      return [
-        'www' => [
-          '#title' => $this->get('name'),
-          '#type' => 'link',
-          '#url' => Url::fromUri($link),
-        ],
-      ];
-    }
-
-    return $build;
-  }
 
 }
