+++ src/Field/Connection/OpeningHourObject.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\helfi_tpr\Field\Connection;
+
+/**
+ * Provides a DTO for TPR connection type of OPENING_HOUR_OBJECT.
+ */
+final class OpeningHourObject extends OpeningHourBase {
+
+  /**
+   * The type name.
+   *
+   * @var string
+   */
+  public const TYPE_NAME = 'OPENING_HOUR_OBJECT';
+
+}
